Add page consumption to cross-links command
The issues cross-links command previously did not take into account the
fact that JQL Search results are paged so would only find the cross
links for the next 50 issues (the default page size for the relevant
REST API call).  The command now continues to loop making a query to get
the next page of issues until there are no further issues.